Rename filenames in #include commands to match real filenames
