+++ src/Native/Runtime/DebugEventSource.cpp
@@ -7,20 +7,20 @@
 #include "gcrhenv.h"
 #endif // DACCESS_COMPILE
 #if !defined(DACCESS_COMPILE)
-#include "commontypes.h"
+#include "CommonTypes.h"
 #include "daccess.h"
-#include "commonmacros.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "CommonMacros.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "assert.h"
 #include "static_check.h"
 #include "type_traits.hpp"
 #include "slist.h"
 #include "holder.h"
-#include "crst.h"
-#include "instancestore.h"
-#include "rwlock.h"
-#include "runtimeinstance.h"
+#include "Crst.h"
+#include "InstanceStore.h"
+#include "RWLock.h"
+#include "RuntimeInstance.h"
 #include "gcrhinterface.h"
 #include "module.h"
 #include "DebugEventSource.h"
