+++ src/Native/Runtime/EHHelpers.cpp
@@ -4,26 +4,26 @@
 //
 #ifndef DACCESS_COMPILE
 #include "rhcommon.h"
-#include "commontypes.h"
+#include "CommonTypes.h"
 #include "daccess.h"
-#include "commonmacros.h"
+#include "CommonMacros.h"
 #include "assert.h"
 #include "slist.h"
 #include "gcrhinterface.h"
 #include "module.h"
 #include "varint.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "holder.h"
-#include "crst.h"
-#include "rwlock.h"
-#include "runtimeinstance.h"
+#include "Crst.h"
+#include "RWLock.h"
+#include "RuntimeInstance.h"
 #include "event.h"
 #include "threadstore.h"
 #include "regdisplay.h"
-#include "stackframeiterator.h"
+#include "StackFrameIterator.h"
 #include "thread.h"
-#include "stresslog.h"
+#include "stressLog.h"
 
 // Find the module containing the given address, which is a return address from a managed function. The
 // address may be to another managed function, or it may be to an unmanaged function, or it may be to a GC
