+++ src/Native/Runtime/MiscHelpers.cpp
@@ -11,34 +11,34 @@
 #ifdef DACCESS_COMPILE
 #include "gcrhenv.h"
 #endif // DACCESS_COMPILE
-#include "commontypes.h"
+#include "CommonTypes.h"
 #include "daccess.h"
-#include "commonmacros.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "CommonMacros.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "assert.h"
 #include "static_check.h"
 #include "slist.h"
 #include "holder.h"
-#include "crst.h"
+#include "Crst.h"
 #include "rhbinder.h"
-#include "rwlock.h"
-#include "runtimeinstance.h"
+#include "RWLock.h"
+#include "RuntimeInstance.h"
 #include "regdisplay.h"
 #include "gcrhinterface.h"
 #include "varint.h"
-#include "stackframeiterator.h"
+#include "StackFrameIterator.h"
 #include "thread.h"
 #include "event.h"
 #include "threadstore.h"
 #include "gcrhinterface.h"
 #include "module.h"
 #include "eetype.h"
-#include "objectlayout.h"
-#include "genericinstance.h"
+#include "ObjectLayout.h"
+#include "GenericInstance.h"
 #include "slist.inl"
 #include "eetype.inl"
-#include "commonmacros.inl"
+#include "CommonMacros.inl"
 
 // Busy spin for the given number of iterations.
 COOP_PINVOKE_HELPER(void, RhSpinWait, (Int32 iterations))
