+++ src/Native/Runtime/RestrictedCallouts.cpp
@@ -9,29 +9,29 @@
 //
 
 #include "rhcommon.h"
-#include "commontypes.h"
-#include "commonmacros.h"
+#include "CommonTypes.h"
+#include "CommonMacros.h"
 #include "daccess.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "assert.h"
 #include "slist.h"
 #include "holder.h"
 #include "gcrhinterface.h"
 #include "module.h"
 #include "rhbinder.h"
-#include "crst.h"
-#include "rwlock.h"
-#include "runtimeinstance.h"
+#include "Crst.h"
+#include "RWLock.h"
+#include "RuntimeInstance.h"
 #include "eetype.h"
-#include "objectlayout.h"
+#include "ObjectLayout.h"
 #include "event.h"
 #include "varint.h"
 #include "regdisplay.h"
-#include "stackframeiterator.h"
+#include "StackFrameIterator.h"
 #include "thread.h"
 #include "threadstore.h"
-#include "restrictedcallouts.h"
+#include "RestrictedCallouts.h"
 
 // The head of the chains of GC callouts, one per callout type.
 RestrictedCallouts::GcRestrictedCallout * RestrictedCallouts::s_rgGcRestrictedCallouts[GCRC_Count] = { 0 };
