+++ src/Native/Runtime/dllmain.cpp
@@ -3,21 +3,21 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 #include "rhcommon.h"
-#include "commontypes.h"
+#include "CommonTypes.h"
 #include "daccess.h"
-#include "commonmacros.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "CommonMacros.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "gcrhinterface.h"
 
 #include "assert.h"
 #include "slist.h"
 #include "varint.h"
 #include "regdisplay.h"
-#include "stackframeiterator.h"
+#include "StackFrameIterator.h"
 #include "thread.h"
 #include "holder.h"
-#include "crst.h"
+#include "Crst.h"
 #include "event.h"
 
 bool InitDLL(HANDLE hPalInstance);
