+++ src/Native/Runtime/eetype.cpp
@@ -8,14 +8,14 @@
 #endif // DACCESS_COMPILE
 
 #ifndef DACCESS_COMPILE
-#include "commontypes.h"
+#include "CommonTypes.h"
 #include "daccess.h"
-#include "commonmacros.h"
+#include "CommonMacros.h"
 #include "assert.h"
 #include "rhbinder.h"
 #include "eetype.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #endif
 
 #pragma warning(disable:4127) // C4127: conditional expression is constant
