+++ src/Native/Runtime/eetype.h
@@ -636,4 +636,4 @@ public:
 
 #pragma warning(pop)
 
-#include "optionalfields.h"
+#include "OptionalFields.h"
