+++ src/Native/Runtime/startup.cpp
@@ -8,30 +8,30 @@
 #endif // DACCESS_COMPILE
 
 #ifndef DACCESS_COMPILE
-#include "commontypes.h"
+#include "CommonTypes.h"
 #include "daccess.h"
-#include "commonmacros.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "CommonMacros.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "assert.h"
 #include "slist.h"
 #include "gcrhinterface.h"
 #include "varint.h"
 #include "regdisplay.h"
-#include "stackframeiterator.h"
+#include "StackFrameIterator.h"
 #include "thread.h"
 #include "holder.h"
-#include "crst.h"
+#include "Crst.h"
 #include "event.h"
-#include "rwlock.h"
+#include "RWLock.h"
 #include "threadstore.h"
-#include "runtimeinstance.h"
-#include "instancestore.h"
+#include "RuntimeInstance.h"
+#include "InstanceStore.h"
 #include "rhbinder.h"
-#include "cachedinterfacedispatch.h"
-#include "rhconfig.h"
-#include "stresslog.h"
-#include "restrictedcallouts.h"
+#include "CachedInterfaceDispatch.h"
+#include "RhConfig.h"
+#include "stressLog.h"
+#include "RestrictedCallouts.h"
 #endif !DACCESS_COMPILE
 
 #ifndef DACCESS_COMPILE
