+++ src/Native/Runtime/stressLog.cpp
@@ -15,22 +15,22 @@
 #endif // DACCESS_COMPILE
 
 #ifndef DACCESS_COMPILE
-#include "commontypes.h"
-#include "commonmacros.h"
-#include "palredhawkcommon.h"
-#include "palredhawk.h"
+#include "CommonTypes.h"
+#include "CommonMacros.h"
+#include "PalRedhawkCommon.h"
+#include "PalRedhawk.h"
 #include "daccess.h"
-#include "stresslog.h"
+#include "stressLog.h"
 #include "holder.h"
-#include "crst.h"
+#include "Crst.h"
 #include "assert.h"
 #include "slist.h"
 #include "gcrhinterface.h"
 #include "varint.h"
 #include "regdisplay.h"
-#include "stackframeiterator.h"
+#include "StackFrameIterator.h"
 #include "thread.h"
-#include "rwlock.h"
+#include "RWLock.h"
 #include "event.h"
 #include "threadstore.h"
